feat: support azure tenant and guest users
This commit allows the user to set up the endpoint (if using
mooncake) and also incorporates the files for supporting tenant
configuration on azure.

Introduces as well a condition for setting the user id, used for
cases where the users don't have an upn (i.e. guest users on a
private tenant)